requirements.txt
flask
tensorflow>=2.12
numpy
pillow
matplotlib
plotly
folium
geopandas
rasterio
scikit-learn
gunicorn